Detail排版修改

+++ RouteMaster/Controllers/AccommodationsController.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
+using System.Web.Security;
 using Antlr.Runtime;
 using Microsoft.Ajax.Utilities;
 using Newtonsoft.Json.Linq;
@@ -27,8 +28,8 @@ using static RouteMaster.Filter.PartnerAuthenticationFilter;
 
 namespace RouteMaster.Controllers
 {
-    //[PartnerAuthenticationFilter]
-    //[PartnerAuthorizeAttribute]
+    [PartnerAuthenticationFilter]
+    [PartnerAuthorizeAttribute]
 
     public class AccommodationsController : Controller
     {
@@ -42,17 +43,51 @@ namespace RouteMaster.Controllers
         }
         
         // 合夥人的住宿列表
-        public ActionResult MyAccommodationIndex(int? id)
+        public ActionResult MyAccommodationIndex()
         {
-			IEnumerable<AccommodationIndexVM> accommodations = GetAccommodations(id);
+            var profile = GetPartnerEmailAndId();
+			// 取得 HTTP 請求中的 Cookie 集合
+			if( !string.IsNullOrEmpty(profile.Item1) && profile.Item2 != 0)
+            {
+			    IEnumerable<AccommodationIndexVM> accommodations = GetAccommodations(profile.Item2);
             
-            //var accommodations = db.Accommodations.Include(a => a.Partner).Include(a => a.Region).Include(a => a.Town);
+                //var accommodations = db.Accommodations.Include(a => a.Partner).Include(a => a.Region).Include(a => a.Town);
 
-            return View(accommodations);//.ToList());
+                return View(accommodations);//.ToList());
+            }
+
+            return RedirectToAction("PartnerLogin", "Partners");
         }
 
-        // 還沒做
-        public ActionResult Details(int? id)
+		private (string, int) GetPartnerEmailAndId()
+		{
+			HttpCookieCollection cookies = Request.Cookies;
+
+			// 檢查是否存在特定名稱的 Cookie
+			if (cookies[FormsAuthentication.FormsCookieName] != null)
+			{
+				// 從 Cookie 中取得票據的值
+				string encryptedTicket = cookies[FormsAuthentication.FormsCookieName].Value;
+
+				// 解密票據
+				var ticket = FormsAuthentication.Decrypt(encryptedTicket);
+
+				// 檢查票據是否成功解密
+				if (ticket != null && ticket.Expired == false)
+				{
+					// 取得票據中的使用者資料
+					string email = ticket.Name;
+					int id = int.Parse(ticket.UserData);
+
+                    return (email, id);
+				}
+                return (null, 0);
+			}
+            return(null, 0);
+		}
+
+		// 還沒做
+		public ActionResult Details(int? id)
         {
             if (id == null)
             {
@@ -151,6 +186,14 @@ namespace RouteMaster.Controllers
 			
         }
 
+        public ActionResult RoomsIndex(int id)
+        {
+            var rooms = db.Rooms.Where(r => r.AccommodationId == id).Include(a => a.RoomImages);
+
+            ViewBag.Id = id;
+            return View(rooms.ToList().Select(r => r.ToVM()));
+        }
+
 		// 新增客房
 		public ActionResult CreateRoom(int? id)
         {
@@ -194,8 +237,8 @@ namespace RouteMaster.Controllers
 			}
 			else
 			{
-				ModelState.AddModelError(string.Empty, result.ErrorMessage);
                 PrepareRoomTypeViewBag();
+				ModelState.AddModelError(string.Empty, result.ErrorMessage);
 				return View(vm);
 			}
 			//ViewBag.PartnerId = new SelectList(db.Partners, "Id", "FirstName", accommodation.PartnerId);
@@ -205,14 +248,48 @@ namespace RouteMaster.Controllers
 		}
 
 		// 客房列表
-        public ActionResult RoomsIndex(int id)
-        {
-            var rooms = db.Rooms.Where(r => r.AccommodationId == id).Include(a => a.RoomImages);
+		public ActionResult EditRoom(int? id)
+		{
+			if (id == null)
+			{
+				return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+			}
 
-            ViewBag.Id = id;
-            return View(rooms.ToList().Select(r => r.ToVM()));
-        }
+			//var model = GetMemberProfile(currentUserAccount);
+			RoomEditVM vm = GetRoomProfile(id);
 
+			//Accommodation accommodation = db.Accommodations.Find(id);
+			if (vm == null)
+			{
+				return HttpNotFound();
+			}
+
+			PrepareRoomTypeViewBag();
+
+			return View(vm);
+		}
+
+		[HttpPost]
+		[ValidateAntiForgeryToken]
+		public ActionResult EditRoom(RoomEditVM vm, ImagesVM iVM)
+		{
+			if (!ModelState.IsValid) return View(vm);
+
+			Result result = EditRoomProfile(vm, iVM);
+
+			if (result.IsSuccess)
+			{
+
+				return RedirectToAction("RoomsIndex", new { id = vm.AccommodationId });
+			}
+			else
+			{
+                PrepareRoomTypeViewBag();
+				ModelState.AddModelError(string.Empty, result.ErrorMessage);
+				return View(vm);
+			}
+
+		}
 
 		// 編輯公共設施
 		public ActionResult EditServiceInfo(int? id)
@@ -279,21 +356,41 @@ namespace RouteMaster.Controllers
         public void DeleteConfirmed(int id)
         {
             Accommodation accommodation = db.Accommodations.Find(id);
-            foreach(var ai in accommodation.AccommodationImages)
+            var ais = accommodation.AccommodationImages;
+            var rs = accommodation.Rooms;
+            var ss = accommodation.ServiceInfos;
+
+			if (ais != null || ais.Count() != 0)
             {
-               db.AccommodationImages.Remove(ai);
-            }
-            
-            foreach(var r in accommodation.Rooms)
-            {
-               db.Rooms.Remove(r);
-            }
-            
-            foreach(var ris in accommodation.Rooms.Select(r=>r.RoomImages))
-            {
-                foreach(var ri in ris)
+                foreach(var ai in ais.ToList())
                 {
-                    db.RoomImages.Remove(ri);
+                   DeleteUploadFile(ai.Image);
+                   db.AccommodationImages.Remove(ai);
+                }
+            }
+
+
+			if (rs != null || rs.Count() != 0)
+            {
+			    foreach (var r in rs.ToList())
+			    {
+                    var ris = r.RoomImages;
+					if (ris != null || ris.Count() != 0)
+                    {
+				        foreach (var ri in r.RoomImages.ToList())
+				        {
+                            DeleteUploadFile(ri.Image);
+					        db.RoomImages.Remove(ri);
+				        }
+                    }
+                   db.Rooms.Remove(r);
+			    }
+            }
+            if(ss!= null || ss.Count() != 0)
+            {
+                foreach(var s in ss.ToList())
+                {
+					accommodation.ServiceInfos.Remove(s);
                 }
             }
             
@@ -301,8 +398,29 @@ namespace RouteMaster.Controllers
             db.SaveChanges();
         }
 
+		private void DeleteUploadFile(string file1)
+		{
+			string path = Server.MapPath("~/Uploads");
+			string fullName = Path.Combine(path, file1);
+			System.IO.File.Delete(fullName);
+		}
 
+		[HttpPost]
+     
+        public void DeleteRoom(int id)
+        {
+            Room room = db.Rooms.Find(id);
 
+            foreach(var ri in room.RoomImages.ToList())
+            {
+               DeleteUploadFile(ri.Image);
+               db.RoomImages.Remove(ri);
+            }
+
+            db.Rooms.Remove(room);
+            db.SaveChanges();
+
+        }
 
 
 
@@ -338,27 +456,46 @@ namespace RouteMaster.Controllers
 		}
 		private void PrepareRoomTypeViewBag()
 		{
-            var roomTypes = new List<RoomType>{
-                new RoomType(1,"單人房"),
-                new RoomType(2,"雙人房"),
-                new RoomType(3,"雙床房"),
-                new RoomType(4,"三人房"),
-                new RoomType(5,"四人房"),
-                new RoomType(6,"家庭房"),
-                new RoomType(7,"套房"),
-                new RoomType(8,"雅房"),
-                new RoomType(9,"膠囊床位")
+            var roomTypes = new List<SelectListItem>
+            {
+                new SelectListItem { Value = null, Text = "請選擇", Disabled = true, Selected = true },
+                new SelectListItem { Value = "單人房", Text = "單人房" },
+                new SelectListItem { Value = "雙人房", Text = "雙人房" },
+                new SelectListItem { Value = "雙床房", Text = "雙床房" },
+                new SelectListItem { Value = "三人房", Text = "三人房" },
+                new SelectListItem { Value = "四人房", Text = "四人房" },
+                new SelectListItem { Value = "家庭房", Text = "家庭房" },
+                new SelectListItem { Value = "套房", Text = "套房" },
+                new SelectListItem { Value = "雅房", Text = "雅房" },
+                new SelectListItem { Value = "膠囊床位", Text = "膠囊床位" }
             };
 
-            ViewBag.RoomType = new SelectList(roomTypes, "Type", "Type")
-				.Prepend(new SelectListItem
-				{
-					Disabled = true,
-					Selected = true,
-					Text = "請選擇",
-					Value = null
-				});
-		}
+            ViewBag.RoomType = roomTypes;
+
+            //         var roomTypes = new List<RoomType>{
+            //             new RoomType("單人房","單人房"),
+            //             new RoomType("雙人房","雙人房"),
+            //             new RoomType("雙床房","雙床房"),
+            //             new RoomType("三人房","三人房"),
+            //             new RoomType("四人房","四人房"),
+            //             new RoomType("家庭房","家庭房"),
+            //             new RoomType("套房","套房"),
+            //             new RoomType("雅房","雅房"),
+            //             new RoomType("膠囊床位","膠囊床位")
+            //         };
+
+
+
+            //ViewBag.RoomType = new SelectList(roomTypes.Select(rt => new SelectListItem { Value = rt.Value, Text = rt.Text })
+            //.Prepend(new SelectListItem
+            //{
+            //	Disabled = true,
+            //	Selected = true,
+            //	Text = "請選擇",
+            //	Value = null
+            //}));
+
+        }
 
 		private Result EditAccommodationProfile(AccommodationEditVM vm, ImagesVM iVM)
 		{
@@ -367,6 +504,15 @@ namespace RouteMaster.Controllers
 			AccommodationService service = new AccommodationService(repo);
 
 			return service.EditAccommodationProfile(vm.ToDto(), iVM.ToDto(), path);
+		}
+
+        private Result EditRoomProfile(RoomEditVM vm, ImagesVM iVM)
+		{
+			string path = Server.MapPath("~/Uploads");
+			IAccommodationRepository repo = new AccommodationEFRepository();
+			AccommodationService service = new AccommodationService(repo);
+
+			return service.EditRoomProfile(vm.ToDto(), iVM.ToDto(), path);
 		}
 
         protected override void Dispose(bool disposing)
@@ -384,6 +530,15 @@ namespace RouteMaster.Controllers
 			//IProductRepository repo = new ProductDapperRepository();
 			AccommodationService service = new AccommodationService(repo);
             return service.GetEditInfo(id)?.ToVM();
+
+		}
+        
+        private RoomEditVM GetRoomProfile(int? id)
+		{
+			IAccommodationRepository repo = new AccommodationEFRepository();
+			//IProductRepository repo = new ProductDapperRepository();
+			AccommodationService service = new AccommodationService(repo);
+            return service.GetRoomInfo(id)?.ToVM();
 		}
 
         private Result CreateAccommodation(AccommodationCreateVM vm)
@@ -432,8 +587,18 @@ namespace RouteMaster.Controllers
         [HttpPost]
         public void DeleteImage(string imgPath)
         {
-            AccommodationImage ai = db.AccommodationImages.FirstOrDefault(a=>a.Image ==imgPath);
+            AccommodationImage ai = db.AccommodationImages.FirstOrDefault(a => a.Image == imgPath);
             db.AccommodationImages.Remove(ai);
+
+			db.SaveChanges();
+
+        }
+        
+        [HttpPost]
+        public void DeleteRoomImage(string imgPath)
+        {
+            RoomImage ri = db.RoomImages.FirstOrDefault(r=>r.Image == imgPath);
+            db.RoomImages.Remove(ri);
 
 			db.SaveChanges();
